requirements.txt
Flask
Flask-Login 
Jinja2
WTForms
Flask-WTF
bcrypt
blinker
certifi
cffi
click
Flask-Bcrypt
Flask-Mail
Flask-SQLAlchemy
itsdangerous
flask_mysqldb